Canny Edge Detection in python openCV (Bangla)
In this video i will teach you how to detection Canny Edge images Using Python & OpenCV.
OpenCV provides method called Canny for Canny edge detector.

What is canny image?
The Canny edge detector is an edge detection operator that uses a multi-stage algorithm to detect a wide 
range of edges in images. It was developed by John F. Canny in 1986. Canny also produced a computational 
theory of edge detection explaining why the technique works. It accepts a gray scale image as input and it 
uses a multistage algorithm. You can perform this operation on an image using the Canny() method of the 
imgproc class, following is the syntax of this method.

1: Noise Reduction : Since edge detection is susceptible to noise in the image, first step is to remove the noise in the image with a 5x5 Gaussian filter.

2: Calculate the gradients 

3: Non-maximum Suppression: After getting gradient magnitude and direction, a full scan of image is done to remove any unwanted pixels

which may not constitute the edge. For this, at every pixel, pixel is checked if it is a local maximum in its neighborhood in the direction of gradient

4: Double threshold on all the detected edges 

5: Analyzes all the edges and their connection